move downloading functions
